+++ mokapot/confidence.py
@@ -15,9 +15,9 @@ We recommend using the :py:func:`~mokapot.brew()` function or the
 :py:meth:`~mokapot.LinearPsmDataset.assign_confidence()` method to obtain these
 confidence estimates, rather than initializing the classes below directly.
 """
-import os
 import copy
 import logging
+from pathlib import Path
 
 import pandas as pd
 import matplotlib.pyplot as plt
@@ -26,6 +26,7 @@ from triqler import qvality
 from . import qvalues
 from . import utils
 from .picked_protein import picked_protein, crosslink_picked_protein
+from .writers import to_flashlfq, to_txt
 
 LOGGER = logging.getLogger(__name__)
 
@@ -99,9 +100,15 @@ class GroupedConfidence:
         """The groups for confidence estimation"""
         return list(self._group_confidence_estimates.keys())
 
-    def to_txt(self, dest_dir=None, file_root=None, sep="\t", decoys=False):
-        """
-        Save confidence estimates to delimited text files.
+    def to_txt(
+        self,
+        dest_dir=None,
+        file_root=None,
+        sep="\t",
+        decoys=False,
+        combine=False,
+    ):
+        """Save confidence estimates to delimited text files.
 
         Parameters
         ----------
@@ -109,19 +116,36 @@ class GroupedConfidence:
             The directory in which to save the files. `None` will use the
             current working directory.
         file_root : str or None, optional
-            An optional prefix for the confidence estimate files. The
-            suffix will always be `mokapot.psms.txt` and
-            `mokapot.peptides.txt`.
+            An optional prefix for the confidence estimate files. The suffix
+            will be "mokapot.{level}.txt", where "{level}" indicates the level
+            at which confidence estimation was performed (i.e. PSMs, peptides,
+            proteins) if :code:`combine=True`. If :code:`combine=False` (the
+            default), additionally the group value is prepended, yeilding a
+            suffix "{group}.mokapot.{level}.txt".
         sep : str, optional
             The delimiter to use.
         decoys : bool, optional
             Save decoys confidence estimates as well?
+        combine : bool, optional
+            Should groups be combined into a single file?
 
         Returns
         -------
         list of str
             The paths to the saved files.
+
         """
+        if combine:
+            res = self.group_confidence_estimates.values()
+            ret_files = to_txt(
+                res,
+                dest_dir=dest_dir,
+                file_root=file_root,
+                sep=sep,
+                decoys=decoys,
+            )
+            return ret_files
+
         ret_files = []
         for group, res in self.group_confidence_estimates.items():
             prefix = file_root + f".{group}"
@@ -205,8 +229,7 @@ class Confidence:
         return list(self.confidence_estimates.keys())
 
     def to_txt(self, dest_dir=None, file_root=None, sep="\t", decoys=False):
-        """
-        Save confidence estimates to delimited text files.
+        """Save confidence estimates to delimited text files.
 
         Parameters
         ----------
@@ -214,9 +237,10 @@ class Confidence:
             The directory in which to save the files. `None` will use the
             current working directory.
         file_root : str or None, optional
-            An optional prefix for the confidence estimate files. The
-            suffix will always be `mokapot.psms.txt` and
-            `mokapot.peptides.txt`.
+            An optional prefix for the confidence estimate files. The suffix
+            will always be "mokapot.{level}.txt", where "{level}" indicates the
+            level at which confidence estimation was performed (i.e. PSMs,
+            peptides, proteins).
         sep : str, optional
             The delimiter to use.
         decoys : bool, optional
@@ -226,32 +250,15 @@ class Confidence:
         -------
         list of str
             The paths to the saved files.
+
         """
-        file_base = "mokapot"
-        if file_root is not None:
-            file_base = file_root + "." + file_base
-        if dest_dir is not None:
-            file_base = os.path.join(dest_dir, file_base)
-
-        out_files = []
-        for level, qvals in self.confidence_estimates.items():
-            if qvals is None:
-                continue
-
-            out_file = file_base + f".{level}.txt"
-            qvals.to_csv(out_file, sep=sep, index=False)
-            out_files.append(out_file)
-
-        if decoys:
-            for level, qvals in self.decoy_confidence_estimates.items():
-                if qvals is None:
-                    continue
-
-                out_file = file_base + f".decoys.{level}.txt"
-                qvals.to_csv(out_file, sep=sep, index=False)
-                out_files.append(out_file)
-
-        return out_files
+        return to_txt(
+            self,
+            dest_dir=dest_dir,
+            file_root=file_root,
+            sep=sep,
+            decoys=decoys,
+        )
 
     def _perform_tdc(self, psm_columns):
         """Perform target-decoy competition.
@@ -340,7 +347,6 @@ class LinearConfidence(Confidence):
         A dictionary of confidence estimates at each level.
     decoy_confidence_estimates : Dict[str, pandas.DataFrame]
         A dictionary of confidence estimates for the decoys at each level.
-
     """
 
     def __init__(self, psms, scores, desc=True, eval_fdr=0.01):
@@ -350,6 +356,8 @@ class LinearConfidence(Confidence):
         self._data[self._target_column] = psms.targets
         self._psm_columns = psms._spectrum_columns
         self._peptide_column = psms._peptide_column
+        self._protein_column = psms._protein_column
+        self._optional_columns = psms._optional_columns
         self._eval_fdr = eval_fdr
 
         LOGGER.info("Performing target-decoy competition...")
@@ -455,7 +463,6 @@ class LinearConfidence(Confidence):
                     scores[targets], scores[~targets], includeDecoys=True
                 )
             except SystemExit as msg:
-                print(msg)
                 if "no decoy hits available for PEP calculation" in str(msg):
                     pep = 0
                 else:
@@ -463,12 +470,42 @@ class LinearConfidence(Confidence):
 
             level = level.lower()
             data["mokapot PEP"] = pep
+            if level != "proteins" and self._protein_column is not None:
+                data[self._protein_column] = data.pop(self._protein_column)
+
             self.confidence_estimates[level] = data.loc[targets, :]
             self.decoy_confidence_estimates[level] = data.loc[~targets, :]
 
         if "proteins" not in self.confidence_estimates.keys():
             self.confidence_estimates["proteins"] = None
             self.decoy_confidence_estimates["proteins"] = None
+
+    def to_flashlfq(self, out_file="mokapot.flashlfq.txt"):
+        """Save confidenct peptides for quantification with FlashLFQ.
+
+        `FlashLFQ <https://github.com/smith-chem-wisc/FlashLFQ>`_ is an
+        open-source tool for label-free quantification. For mokapot to save
+        results in a compatible format, a few extra columns are required to
+        be present, which specify the MS data file name, the theoretical
+        peptide monoisotopic mass, the retention time, and the charge for each
+        PSM. If these are not present, saving to the FlashLFQ format is
+        disabled.
+
+        Note that protein grouping in the FlashLFQ results will be more
+        accurate if proteins were added for analysis with mokapot.
+
+        Parameters
+        ----------
+        out_file : str, optional
+            The output file to write.
+
+        Returns
+        -------
+        str
+            The path to the saved file.
+
+        """
+        return to_flashlfq(self, out_file)
 
 
 class CrosslinkConfidence(Confidence):