+++ mokapot/picked_protein.py
@@ -119,8 +119,8 @@ def picked_protein(
         The column in `peptides` containing the peptide sequence.
     score_column: str
         The column in `peptides` containing the scores.
-    proteins : FastaProteins object
-        A FastaProteins object.
+    proteins : Proteins object
+        A Proteins object.
 
     Returns
     -------
@@ -184,7 +184,7 @@ def group_with_decoys(peptides, proteins):
     ----------
     peptides : pandas.DataFrame
         The peptide dataframe.
-    proteins : FastaProteins object
+    proteins : a Proteins object
 
     Returns
     -------
@@ -206,7 +206,7 @@ def group_without_decoys(peptides, target_column, proteins):
         The peptide dataframe.
     target_column : str
         The column indicating if the peptide is a target.
-    proteins : a FastaProteins object
+    proteins : a Proteins object
 
     Returns
     -------